pyproject.toml
[project]
name = "jutulgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.5",
    "flashrank>=0.2.10",
    "ipython>=9.4.0",
    "jutuldarcy>=1.1.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.27",
    "langchain-text-splitters>=0.3.8",
    "langchainhub>=0.1.21",
    "langgraph>=0.5.0",
    "langgraph-api>=0.2.78",
    "langgraph-cli[inmem]>=0.3.3",
    "langgraph-sdk>=0.1.73",
    "langsmith>=0.4.4",
    "pydantic>=2.11.7",
    "pygraphviz>=1.14",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "rich>=14.1.0",
    "ruff>=0.12.4",
    "transformers>=4.53.3",
    "unstructured[md]>=0.18.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jutulgpt"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
